requirements-dev.txt
mypy==0.961
flake8==4.0.1
pytest==7.1.2
factory-boy==3.2.1
